requirements.txt
fastapi
uvicorn
websockets
python-dotenv
torchaudio
webrtcvad
pyannote.audio
openai
scipy
numpy